requirements.txt
# Core dependencies
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
python-pptx>=0.6.21
mutagen>=1.47.0
opencv-python>=4.8.0
python-magic>=0.4.27
pandas>=2.0.0
tqdm>=4.65.0
numpy>=1.24.0

# Web interface
flask>=2.3.0
flask-cors>=4.0.0
jinja2>=3.1.0
werkzeug>=2.3.0

# Database
sqlite3

# Utilities
click>=8.1.0
pyyaml>=6.0
colorama>=0.4.6
rich>=13.0.0
tabulate>=0.9.0

# Security
cryptography>=41.0.0
hashlib

# Development (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0
